Styling overhaul
Added a background and a background to every house depending on what biome is chosen. Slightly cleaned App component. Changed the name of biome icons. Added Biome background images. Disabled spacing on cardaction to let the marginLeft work on placing the collapse arrow on the right side. Added Card Headers to make the cards more compact and to help positioning. Changed Delete button to a IconButton delete. Added a small list of npc likes/dislikes on NPC card. Fixed some of the preferences.json typos. I added a Theme.js called before app.js to set the theme (Might change this soon). I plan to have the backgrounds set as a changable theme as well as adding dark and light theme if users find the backgrounds annoying. (This might not happen but its what I plan on doing)