Added pylons next to biome selection
Set the size of icons manually to make sure it doesn't push things out of place (They look a bit weird now though). Changed from emoji's for the npc preferences to material icons. Chaned the shown happiness from a decimal to a percentage (occsionally the value when multipied by 100 would have a lot of .9999 so just rounded it to the nearest integer after multpication). Added pylons to the right of the biome selection when there are two npcs in the town and when clicked it will represent the pylon being placed in game by turning into a .gif of the pylon.