Modularized priceCalc and added princess npc
Added tooltip to npc price modifier to tell the user how the price was calculated. Added princess icon, some calculations specfically for the princess (not finished she is complicated for the way I have handled npc happiness) and her to the preferences json. Changed the pylon tooltips to represent what the pylon currently represents and gave the pylon a disabled look when lower than two npcs are in the same town.