Added settings such as multi biomes and notes
Modified App class constructor to use a reusable function. Added a function to App.js that prepares data to be calculated in PriceCalc.js. Added multi biome system which allows the user to change a specfic npc's biome. Added menu items to the appbar menu that change whether the user wants multi biomes or notes and More info which is a link to the terraria npc wiki. Changed Biome.js to be reusable for the multi biome npc's. Added Images.js to help remove the clutter of imports and ternarys, now to get the background and pylon images it calls functions from Images.js. PriceCalc has changed to help it become reusable for the multi biomes. Changed description in index.html. Changed private status to true in package.json. Formatted and added new objects in sample.json