Added quick import/export system in menu
added import in which the user has to paste their townsdata in json format into the dialog textfield when they click on the import menuitem. Added export in which the user can download or copy to their clipboard their current towns data through the export menuitem. The import/export system needs to look a bit nicer, it was a quick implementation. Added two npm modules clipboard-copy and file-saver. The functions for TownAppBar.js probably need cleaning.